rename: useExamCodeVerification 훅을 컨벤션에 맞는 위치로 이동 (#66)
