+++ src/hooks/api/index.ts
@@ -1,3 +1,4 @@
 import useExamList from "@/hooks/api/useExamList";
+import useExamCodeVerification from "@/hooks/api/useExamCodeVerification";
 
-export {useExamList}
+export { useExamList, useExamCodeVerification };
