+++ src/hooks/index.ts
@@ -3,7 +3,6 @@ import useWindowSize from "@/hooks/useWindowSize";
 import useOutsideInteraction from "@/hooks/useOutsideInteraction";
 import useModalContext, { ModalContext } from "@/hooks/useModalContext";
 import useInfiniteScroll from "@/hooks/useInfiniteScroll";
-import useExamCodeVerification from "@/hooks/useExamCodeVerification";
 
 export {
   useToast,
@@ -12,5 +11,4 @@ export {
   useModalContext,
   ModalContext,
   useInfiniteScroll,
-  useExamCodeVerification,
 };
